feat: move language and templating to misc
