+++ translatable_proc/src/macro_generation/translation.rs
@@ -2,8 +2,8 @@ use proc_macro2::TokenStream as TokenStream2;
 use quote::{ToTokens, quote};
 use thiserror::Error;
 use translatable_shared::handle_macro_result;
-use translatable_shared::language::Language;
 use translatable_shared::macros::collections::map_to_tokens;
+use translatable_shared::misc::language::Language;
 
 use crate::data::translations::load_translations;
 use crate::macro_input::input_type::InputType;
@@ -54,7 +54,7 @@ pub fn translation_macro(input: TranslationMacroArgs) -> TokenStream2 {
     let language = match input.language() {
         InputType::Static(language) => language.clone().to_token_stream(),
         InputType::Dynamic(language) => quote! {
-            translatable::shared::language::Language::from(#language)
+            translatable::shared::misc::language::Language::from(#language)
         },
     };
