+++ translatable_shared/src/lib.rs
@@ -1,3 +1,3 @@
-pub mod language;
 pub mod macros;
+pub mod misc;
 pub mod translations;
