+++ translatable_shared/src/macros/mod.rs
@@ -1,3 +1,2 @@
 pub mod collections;
 pub mod errors;
-pub mod templating;
